Fixed major bug in eddy viscosity where a slope of a slope was taken when only the slope was needed.  (Thank you Joaquin for finding this).
The variable ze_x_in is already a slope, and yet in one location the
tan(ze_x_in) is taken which is basically the slope of a slope.

In other locations the variable 'tanphi' is defined as the max ze_x_in
over a wave and then later tan(tanphi) is used to determine if the wave
is indeed breaking.  Again a slope of a slope is used.

This version fixes these issues.